add class Child,Salary,Spouse,Working(bad_smell2)
