+++ src/lib/Spouse.java
@@ -1,0 +1,18 @@
+package lib;
+
+public class Spouse extends Employee {
+
+    //data Spouse
+	private String spouseName;
+	private String spouseIdNumber;
+
+    public Spouse () {
+        super ();
+    }
+
+    public void setSpouse(String spouseName, String spouseIdNumber) {
+		this.spouseName = spouseName;
+		this.spouseIdNumber = idNumber;
+	}
+    	
+}
